Deletion system: can properly delete clips Almost done with deleting track in scene mode
Still have to:
Optimize stuff (trying to have menu managers and everything in VMX_V64.py and Custom Clip Shot file for publicizing _clip_shot) (For the buttons binding, have to find how to unbind)

+++ SpecialSessionComponent.py
@@ -20,7 +20,6 @@ class SpecialSessionComponent(SessionComponent):
         self._alt0_igniter = buttons[int(sqrt(len(buttons))) * -1]
         self._alt1_igniter = buttons[int(sqrt(len(buttons))) * -1 + 1]
         # self._igniter.is_momentary = False
-        self._deletion_bind_functions = [[]] * 4
         self._buttons = buttons
         self._mixer = mixer
         self._transport = transport
@@ -57,21 +56,10 @@ class SpecialSessionComponent(SessionComponent):
             scene.name = 'Scene_' + str(scene_index)
             scene.set_triggered_value(2)
             for track_index in range(self.num_tracks):
-                button = self.clip_launch_buttons[scene_index][track_index]
-                clip_slot = scene.clip_slot(track_index)
-                clip_slot.name = str(track_index) + '_Clip_Slot_' + str(scene_index)
-                deletion_lambda = lambda value, cs=clip_slot: self._delete_clip(cs, value)
-                button.add_value_listener(deletion_lambda)
-                self._deletion_bind_functions[scene_index].append(deletion_lambda)
-
-    def unbind_clip_delete(self):
-        for scene_index in range(self.num_scenes):
-            for track_index in range(self.num_tracks):
-                self.clip_launch_buttons[scene_index][track_index].remove_value_listener(self._deletion_bind_functions[scene_index][track_index])
-        self._deletion_bind_functions = [[]] * 4
+                self.clip_launch_buttons[scene_index][track_index].add_value_listener(lambda value, cs=scene.clip_slot(track_index): self._delete_clip(cs, value))
 
     def _delete_clip(self, clip_slot, value):
-        if clip_slot._clip_slot is not None and clip_slot._clip_slot.has_clip:
+        if self.delete_button.is_pressed() and not self._alt0_igniter.is_pressed() and not self._alt1_igniter.is_pressed() and clip_slot._clip_slot is not None and clip_slot._clip_slot.has_clip:
             clip_slot._clip_slot.delete_clip()
 
     def disconnect(self):
@@ -126,14 +114,13 @@ class SpecialSessionComponent(SessionComponent):
     def _engage_sceneLaunch(self, value):
         if value == 127 and not self._alt0_igniter.is_pressed() and not self.delete_button.is_pressed():  # Scene launch mod open
             self.unbind_clip_launch()
-            for scene_index in range(self.num_scenes - 1):
+            for scene_index in range(self.num_scenes):
                 scene = self.scene(scene_index)
                 scene.name = 'Scene_' + str(scene_index)
                 scene.set_launch_button(self.clip_launch_buttons[scene_index][
                                             -1])  # Button is in push mode while it's toggle (problem comes from scene scripts (have to figure out a way to bypass the problem))
                 scene.set_triggered_value(2)
-            self.set_stop_track_clip_buttons(
-                [self.clip_launch_buttons[-1][track_index] for track_index in range(self.num_tracks - 1)])
+            self.set_stop_track_clip_buttons([self.clip_launch_buttons[-1][track_index] for track_index in range(self.num_tracks - 1)])
         elif value == 0:
             for scene_index in range(self.num_scenes):
                 self.scene(scene_index).set_launch_button(None)
@@ -158,12 +145,11 @@ class SpecialSessionComponent(SessionComponent):
 
     def deletion_manager(self):
         self.delete_button.add_value_listener(self._deletion)
+        self.setup_clip_delete()
         self.update()
 
     def _deletion(self, value):
         if value == 127 and not self._alt0_igniter.is_pressed() and not self._alt1_igniter.is_pressed():
             self.unbind_clip_launch()
-            self.setup_clip_delete()
         if value == 0:
-            self.unbind_clip_delete()
             self.setup_clip_launch()
